Put original scripts back in cookieFunctions and just added print function into site.js because although same as admin portal it threw error in cyress and I cannot find a solution for that.
